+++ packages/amplify-category-custom/CHANGELOG.md
@@ -3,7 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [3.1.26-beta-latest.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-custom@3.1.25...@aws-amplify/amplify-category-custom@3.1.26-beta-latest.0) (2025-02-12)
+## [3.1.27](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-custom@3.1.26...@aws-amplify/amplify-category-custom@3.1.27) (2025-02-20)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-custom
+
+
+
+
+
+## [3.1.26](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-custom@3.1.25...@aws-amplify/amplify-category-custom@3.1.26) (2025-02-07)
 
 
 ### Bug Fixes