+++ packages/amplify-console-integration-tests/CHANGELOG.md
@@ -3,7 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [2.11.20-beta-latest.0](https://github.com/aws-amplify/amplify-console-integration-tests/compare/@aws-amplify/amplify-console-integration-tests@2.11.19...@aws-amplify/amplify-console-integration-tests@2.11.20-beta-latest.0) (2025-02-12)
+## [2.11.21](https://github.com/aws-amplify/amplify-console-integration-tests/compare/@aws-amplify/amplify-console-integration-tests@2.11.20...@aws-amplify/amplify-console-integration-tests@2.11.21) (2025-02-20)
+
+**Note:** Version bump only for package @aws-amplify/amplify-console-integration-tests
+
+
+
+
+
+## [2.11.20](https://github.com/aws-amplify/amplify-console-integration-tests/compare/@aws-amplify/amplify-console-integration-tests@2.11.19...@aws-amplify/amplify-console-integration-tests@2.11.20) (2025-02-07)
 
 **Note:** Version bump only for package @aws-amplify/amplify-console-integration-tests