+++ packages/amplify-util-uibuilder/CHANGELOG.md
@@ -3,7 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.14.17-beta-latest.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.14.16...@aws-amplify/amplify-util-uibuilder@1.14.17-beta-latest.0) (2025-02-12)
+## [1.14.18](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.14.17...@aws-amplify/amplify-util-uibuilder@1.14.18) (2025-02-20)
+
+**Note:** Version bump only for package @aws-amplify/amplify-util-uibuilder
+
+
+
+
+
+## [1.14.17](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.14.16...@aws-amplify/amplify-util-uibuilder@1.14.17) (2025-02-07)
 
 **Note:** Version bump only for package @aws-amplify/amplify-util-uibuilder